backend/requirements.txt
asgiref==3.7.2
astroid==3.1.0
cfgv==3.4.0
dill==0.3.8
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==23.3.0
filelock==3.13.1
flake8==7.0.0
gunicorn==21.2.0
identify==2.5.35
mccabe==0.7.0
nodeenv==1.8.0
packaging==23.2
platformdirs==4.2.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.4
typing_extensions==4.10.0
virtualenv==20.25.1
